+++ harfbuzz/src/hb-ot-shape-complex-default.cc
@@ -96,6 +96,24 @@ const hb_ot_complex_shaper_t _hb_ot_complex_shaper_indic = {
     false, /* fallback_position */
 };
 
+const hb_ot_complex_shaper_t _hb_ot_complex_shaper_khmer = {
+    hb_ot_complex_collect_features_khmer,
+    hb_ot_complex_override_features_khmer,
+    hb_ot_complex_data_create_khmer,
+    hb_ot_complex_data_destroy_khmer,
+    nullptr, /* preprocess_text */
+    nullptr, /* postprocess_glyphs */
+    HB_OT_SHAPE_NORMALIZATION_MODE_COMPOSED_DIACRITICS_NO_SHORT_CIRCUIT,
+    hb_ot_complex_decompose_khmer,
+    hb_ot_complex_compose_khmer,
+    hb_ot_complex_setup_masks_khmer,
+    HB_TAG_NONE, /* gpos_tag */
+    nullptr,     /* reorder_marks */
+    HB_OT_SHAPE_ZERO_WIDTH_MARKS_NONE,
+    false, /* fallback_position */
+};
+
+
 const hb_ot_complex_shaper_t _hb_ot_complex_shaper_thai = {
     nullptr, /* collect_features */
     nullptr, /* override_features */