+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Ported
 - Tables: `CBDT`, `CFF`, `CFF2`, `HVAR`, `MVAR`, `OS/2`, `SVG`, `VORG`, `VVAR`,
   `avar`, `cmap`, `fvar`, `glyf`, `gvar`, `hhea`, `hmtx`, `post`, `sbix`, `vhea`, `vmtx`.
-- Shapers: arabic, hangul, hebrew, indic, thai.
+- Shapers: arabic, hangul, hebrew, indic, khmer, thai.
 - `hb_script_t`
 - `hb_feature_t`
 - `hb_variation_t`